tests/dcase2016/DCASE2016-baseline-system-python-1.0.2/requirements.txt
scipy==0.15.1
numpy==1.9.2
scikit-learn==0.16.1
pyyaml==3.11
librosa==0.4.0
